chore: Resuable `<SearchBar />` component
